requirements.txt
requests
aiohttp
customtkinter
phonenumbers
email-validator
dnspython
beautifulsoup4
yt-dlp